// frontend/src/app/api/model/serviceCatalogViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceCatalogViewDto { 
    id?: number;
    name?: string;
    pricingType?: ServiceCatalogViewDto.PricingTypeEnum;
    pricingValue?: number;
}
export namespace ServiceCatalogViewDto {
    export const PricingTypeEnum = {
        FixedAmount: 'FIXED_AMOUNT',
        PercentageOfBase: 'PERCENTAGE_OF_BASE'
    } as const;
    export type PricingTypeEnum = typeof PricingTypeEnum[keyof typeof PricingTypeEnum];
}