// frontend/src/app/api/model/shipmentUpdateDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {AddressDetailsDto} from './addressDetailsDto';


export interface ShipmentUpdateDto {
    type?: ShipmentUpdateDto.TypeEnum;
    weight?: number;
    length?: number;
    width?: number;
    height?: number;
    paidBy?: ShipmentUpdateDto.PaidByEnum;
    isPaid?: boolean;
    receiverName?: string;
    receiverPhone?: string;
    receiverEmail?: string;
    selectedServiceIds?: Set<number>;
    deliveryOfficeId?: number;
    deliveryAddress?: AddressDetailsDto;
}
export namespace ShipmentUpdateDto {
    export const TypeEnum = {
        Document: 'DOCUMENT',
        Parcel: 'PARCEL',
        Pallet: 'PALLET',
        Oversized: 'OVERSIZED'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const PaidByEnum = {
        Sender: 'SENDER',
        Receiver: 'RECEIVER'
    } as const;
    export type PaidByEnum = typeof PaidByEnum[keyof typeof PaidByEnum];
}